poplar: increase size reserved for FIP
Increase the size that's reserved in the loader image for the FIP to
766 KB (784384 bytes).  This should be enough space to incorporate
UEFI rather than U-Boot as the BL33 image located in the FIP
embedded in "l-loader.bin".  It could be made larger if necessary,
up to 0x1afc00 = 1768448 bytes.  The size selected here arbitarily
limits it so l-loader fits under the 1MB boundary in memory.

Signed-off-by: Alex Elder <elder@linaro.org>