poplar: increase size reserved for FIP again
Increase the size that's reserved in the loader image for the FIP to
0xC0E00 (790016 bytes).  At the time this commit was made, this was
the absolute maximum size this could be and still support a working
boot. Any larger, and the boot would fail, looking like this:

    Entry usb bootstrap

    Usb disk(1):
    Boot head format invalid.

It's possible that the "l-loader" boot code could be rearranged
slightly and allow for a slightly larger working boot image.  This
size is currently adequate to hold UEFI (which is a little over
700KB), but this doesn't leave a lot of cushion for expansion.

Signed-off-by: Alex Elder <elder@linaro.org>